# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Product Listing API
  version: 1.0.0
  description: API for listing available products with search, filtering, sorting, and pagination.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/products:
    get:
      summary: List available products
      description: Returns products that are currently available. Supports search, category filter, sorting, and pagination.
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive partial match against product name.
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          description: Filter products by category.
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          required: false
          description: Sort order.
          schema:
            type: string
            enum: [name_asc, price_asc]
            default: name_asc
        - name: page
          in: query
          required: false
          description: 1-based page number.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Number of items per page.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: Paged list of available products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users:
    post:
      operationId: createUser
      tags:
        - Users
      summary: Create a new user account
      description: Creates a new user account with the provided email, password, and name.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/categories:
    get:
      summary: List categories
      description: Returns categories usable for filtering.
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      required: [id, name, isAvailable]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        descriptionSummary:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        price:
          $ref: '#/components/schemas/Money'
        isAvailable:
          type: boolean
        category:
          $ref: '#/components/schemas/CategoryRef'
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
    CategoryRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Money:
      type: object
      nullable: true
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code.
    ProductListResponse:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    CreateUserRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    UserResponse:
      type: object
      required: [id, email, name, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, PENDING]
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            correlationId:
              type: string
              nullable: true
            details:
              type: array
              nullable: true
              items:
                $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string